+++ src/main/java/racingcar/dao/CarDao.java
@@ -1,0 +1,36 @@
+package racingcar.dao;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.util.List;
+import org.springframework.jdbc.core.BatchPreparedStatementSetter;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+import racingcar.domain.Car;
+
+@Repository
+public class CarDao {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public CarDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public void save(final long savedId, final List<Car> cars) {
+        jdbcTemplate.batchUpdate("INSERT INTO car (play_result_id, name, position) VALUES (?, ?, ?)",
+                new BatchPreparedStatementSetter() {
+                    @Override
+                    public void setValues(final PreparedStatement ps, final int i) throws SQLException {
+                        ps.setLong(1, savedId);
+                        ps.setString(2, cars.get(i).getName());
+                        ps.setInt(3, cars.get(i).getPosition());
+                    }
+
+                    @Override
+                    public int getBatchSize() {
+                        return cars.size();
+                    }
+                });
+    }
+}
